Creates meal plans for Scout Camps
Campfire Butler allows Scout section leaders to automate the process of
creating a meal plan for scout camps.
Campfire Butler calculates the number of meals required for x people
and generates a shopping list.

The meals of a 24hr day assume:
- breakfast
- morning tea
- lunch
- afternoon tea
- dinner & dessert
- supper.

Dinner plans are based off submitted recipes and are offered in a
printable format.